actions: Split out zerver.actions.submessage.
(cherry picked from commit 3a135b04d9e9f84aa2a31d6fc0b1b08e9cf9aeac)
