+++ zerver/tests/test_events.py
@@ -13,6 +13,7 @@ from unittest import mock
 import orjson
 from django.utils.timezone import now as timezone_now
 
+from zerver.actions.submessage import do_add_submessage
 from zerver.actions.typing import check_send_typing_notification, do_send_stream_typing_notification
 from zerver.actions.user_groups import (
     bulk_add_members_to_user_group,
@@ -34,7 +35,6 @@ from zerver.lib.actions import (
     do_add_realm_domain,
     do_add_realm_playground,
     do_add_streams_to_default_stream_group,
-    do_add_submessage,
     do_change_avatar_fields,
     do_change_bot_owner,
     do_change_default_all_public_streams,
